Add files via upload
FIxed the database issue with my register and login page. I attempted to try MongoDB in the past but my macbook repeatedly claimed that its too old for this version of MongoDB regardless of which version I chose to download. Downloading it manually also failed. Because of this, I scrapped MongoDB and attempted SQLite. This is my first time using it so I had help from YouTube and AI to aid me in using it. I learnt quickly and managed to correctly use it. I also added a password strength level in real-time in order to enhance security for my encrypt/decrypt site. I incorporated an email option as well as its validation for further security reasons. Dashboard remains the same and will be aimed to be fixed and enhanced next.